[sweeping] added locator for image plots
